Rename of File and some fixes
I renamed the file funWithLua to PixelRain because I will be keeping all projects in separate folders from now on to keep things more organized. I also fixed a small issue with the audio playback. isStopped changed to isPlaying due to an update in LOVE2d isStopped is no longer a function.